fix(lcov): review follow-ups — u64 line keys, CRLF annotation, bisection
- Widen line keys to u64 end-to-end (FileCoverage, lexer, JaCoCo):
  deletes the u32::try_from panic point in the LCOV glue; spec §3 line
  bound updated from 2^32 to u64, requirement TOML and snapshot match.
  Boundary test flipped: 2^32 parses, 2^64 rejects.
- Move the CRLF MUST annotation off dead code: BufRead::lines already
  strips CRLF; the strip_suffix('\r') line was unreachable for CRLF
  input and is deleted.
- Replace the linear insertion scan in aggregate_da_records with binary
  search discharged against the same lines_sorted invariant; ascending
  producer input is now O(n log n) instead of the worst case.

Verified: cargo verus verify 57/0, 358 tests pass, duvet report --ci
passes with regenerated snapshot.

+++ duvet/src/query/coverage.rs
@@ -36,8 +36,8 @@ impl GenericCoverageData {
 /// Coverage data for a single file
 #[derive(Clone, Debug)]
 pub struct FileCoverage {
-    pub lines: BTreeMap<u32, u64>,          // line_number -> hit_count
-    pub branches: BTreeMap<u32, Vec<bool>>, // line_number -> [taken, not_taken, ...]
+    pub lines: BTreeMap<u64, u64>,          // line_number -> hit_count
+    pub branches: BTreeMap<u64, Vec<bool>>, // line_number -> [taken, not_taken, ...]
 }
 
 impl FileCoverage {
@@ -71,7 +71,7 @@ impl FileCoverage {
             } else {
                 CoverageStatus::Miss
             };
-            record(line_num as u64, status);
+            record(line_num, status);
         }
         for (&line_num, branches) in &self.branches {
             let status = if branches.iter().any(|&taken| taken) {
@@ -79,7 +79,7 @@ impl FileCoverage {
             } else {
                 CoverageStatus::Miss
             };
-            record(line_num as u64, status);
+            record(line_num, status);
         }
         report
     }
